requirements.txt
pydantic>=2.0.0
rich>=13.0.0
python-dotenv>=1.0.0
litellm>=1.0.0
httpx>=0.25.0
sqlite3
